pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "metagenome-generator"
version = "0.1.0"
description = "Download bacterial and viral genomes from NCBI and build simulated metagenome FASTAs for training classifiers."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"
authors = [
    { name = "MetagenomeGenerator contributors" }
]
keywords = ["metagenome", "NCBI", "genomics", "simulation", "training-data"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "biopython>=1.83",
]

[project.optional-dependencies]
dev = [
    "pytest>=7",
    "ruff>=0.1",
]

[project.scripts]
metagenome-generator = "metagenome_generator.cli:main"

[project.urls]
Repository = "https://github.com/yourusername/metagenome-generator"
Documentation = "https://github.com/yourusername/metagenome-generator#readme"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "I"]
